consolidate rust client examples
