+++ guide/docs/api/master/association_config.mdx
@@ -29,7 +29,7 @@ values={[
 <TabItem value="Rust">
 
 ```rust
-{{#include ../dnp3/examples/master_tcp_client.rs:association_config}}
+{{#include ../dnp3/examples/master.rs:association_config}}
 ```
 
 </TabItem>
