+++ guide/docs/api/master/channel_config.mdx
@@ -32,7 +32,7 @@ values={[
 <TabItem value="Rust">
 
 ```rust
-{{#include ../dnp3/examples/master_tcp_client.rs:master_channel_config}}
+{{#include ../dnp3/examples/master.rs:master_channel_config}}
 ```
 
 </TabItem>
